use quantize together
